+++ test/polyfill.js
@@ -1,9 +1,9 @@
 'use strict'
 
-// Polyfill history and fetch api's
-require('html5-history-api')
+// Polyfill fetch api and html5 history api.
 require('es6-promise/auto')
 require('isomorphic-fetch')
+if (window === global) require('html5-history-api')
 
 // Patch jsdom with the fetch api.
 if (!window.Headers) window.Headers = global.Headers