+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 5.0.0 - 2017-05-18
+### Changed
+- Rename 'browser' adapter to 'document' to avoid confusion.
+- Change adapter api to expose { fetch, attach }.
+- Improve documentation.
+
 ## 4.3.1, 4.3.2, 4.3.3 - 2017-03-28
 ### Changed
 - Update dependencies.